JUZI-RAGnet/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-community
langchain-ollama
langchain-chroma
pydantic
python-dotenv
requests
loguru
sentence-transformers